chore: Move FormModal into folder
